internal/cert: check if cert exists before removing
